updated models

+++ src/main/java/com/example/boardmanager/repository/BoardRepository.java
@@ -1,8 +1,7 @@
 package com.example.boardmanager.repository;
 
 import com.example.boardmanager.model.Board;
-import java.math.BigInteger;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardRepository extends JpaRepository<Board, BigInteger> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
 }
